Armature局部播放功能测试1，骨骼/皮肤散架问题需要解决（整体播放也存在）

+++ CC37TmplProject/Code/cocos2d/cocos/platform/CCFileUtils.cpp
@@ -661,7 +661,7 @@ unsigned char* FileUtils::getFileData(const std::string& filename, const char* m
         
 		if (pZFile)
 		{
-			*size = pZFile->size();
+			*size = (ssize_t)pZFile->size();
 		}
 		else
 		{
